+++ addons/sys_components/CfgVehicles.hpp
@@ -1,0 +1,20 @@
+class CfgVehicles {
+    class Man;
+    class CAManBase: Man {
+        acre_antennaMemoryPoints[] = {{"LeftShoulder", "LeftShoulder"}};
+        //acre_antennaMemoryPointsDir[] = {{"Spine3", "Neck"}};
+        acre_antennaDirFnc = QFUNC(getAntennaDirMan);
+    };
+    class Thing;
+    class ACRE_BaseRack : Thing {
+        author = "ACRE2 Team";// TODO: sort out
+        displayName = "ACRE Base Rack";
+        scope = 0;
+        mass = 0;
+        vehicleClass = "";
+        
+        acre_isRack = 1;
+        acre_hasUnique = 1;
+    };
+
+};
