+++ extras/antennas/FA_80.nec
@@ -1,0 +1,14 @@
+CM SEM70 Antenna -- long
+CM 
+CM 
+CM 
+CM 
+CE
+GW	1	9	0	0	0	0	0	1.66	.0001	'Wire 1, 9 segments, halve wavelength long.
+GE	1
+LD	4	1	9	9	50
+GN	1
+EK
+EX	0	1	1	0	1	0	0	'Voltage source (1+j0) at wire 1 segment 1.
+FR	0	55	0	0	30	1
+EN
