+++ extras/antennas/RF_3183_AT0XX.inp
@@ -1,0 +1,17 @@
+CM Vehicle Antenna for 30 to 512MHz (US Military Type; For VRC-103/110/111)
+CM 
+CM 
+CM 
+CM 
+CM forw: 0, 0 ; back: 0, 180
+CE
+GW 1 9 0 0 0 0 0 2.7 .0015
+GE 1
+LD 4 1 9 9 50
+GN 1
+EK
+EX 0 1 1 0 1 0 0
+FR 0 1 0 0 30 0
+XQ
+FR 0 243 0 0 30 2
+RP 0 37 73 1000 -90 0 5 5
