+++ extras/antennas/RF_3183_AT0XX.nec
@@ -1,0 +1,14 @@
+CM Vehicle Antenna for 30 to 512MHz (US Military Type; For VRC-103/110/111)
+CM 
+CM 
+CM 
+CM 
+CE
+GW	1	9	0	0	0	0	0	2.7	.0015	'Wire 1, 9 segments, halve wavelength long.
+GE	1
+LD	4	1	9	9	50
+GN	1
+EK
+EX	0	1	1	0	1	0	0	'Voltage source (1+j0) at wire 1 segment 1.
+FR	0	482	0	0	30	2
+EN
